+++ editor/scaffolds/code/workers/canvas-preview.worker.js
